requirements.txt
coverage==5.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
mysql-connector-python==8.0.20
python-dotenv==0.13.0
redis==3.5.2
